Add epxlicit arguments to codec algorithms.
Previously the algorithms assume that any mentioned interface attribute
is implicitly belonging to the codec interface shared by the method that
invoked the algorithm. Unfortunately, that meant none of the bikeshed
autolinking picked up on usage of the interface attributes within the
algorithm.

Now, all interface attributes are explicitly mentioned when invoking the
algorithm, so the linking picks up on this.

I do think it turns out pretty verbose and hard to read :/. Still
considering other approaches.