Emit VideoDecoderConfig with VideoEncoder output
Supplying the VideoDecoderConfig tells VideoEncoder users how to
eventually decode their encoded chunks. See #108.

The EncodedVideoChunkOutputCallback will emit this config alongside the
EncodedVideoChunk anytime the config changes. State is added to detect
changes and fill out the config.

This also updates the sizing info in VideoEncoderConfig to be more
consistent with VideoDecoderConfig (helpful when generating the decoder
config). See #93.