Migrate spec to Media WG
This update adjusts the metadata of the spec as it gets adopted by the Media
Working Group, and makes the following changes:

- Update CONTRIBUTING.md to match that used in W3C
- Update LICENSE.md to drop the CLA part
- Update links in README.md and to mention Media WG
- Update w3c.json
- Update status, group, and links that currently target the WICG repo in the
spec itself

Also note the switch to `Level: None` in the metadata of the spec as I suspect
that the group will publish the API without any level, at least initially.

Linked to #130.