Adjust registry metadata and build file
The default boilerplate for the Status of This Document section of the Media WG
asserts that the specs are intended to be published as W3C Recommendations. The
text needs to be adjusted for drafts that are intended to be published as W3C
Working Group Notes.

There is no specific way to flag the targeted spec status in Bikeshed. The idea
is thus to add a `Text macro: NOTE yes` macro definition in the spec metadata
and conditionals in the boilerplate files to issue the right text.

The updates to the boilerplate files are at:
https://github.com/tabatkins/bikeshed/pull/1999

This update also makes the jobs that run Bikeshed die on fatal errors and not
on warnings because Bikeshed issues a warning about the lack of a Security and
Privacy section in the registry specs. This warning can be ignored (these
registry specs will probably not get such sections). There are no way to tell
Bikeshed to ignore specific warnings, so all of them have to be.