Migrate spec to Media WG (#150)
This update adjusts the metadata of the spec as it gets adopted by the Media
Working Group, and makes the following changes:

- Update CONTRIBUTING.md to match that used in W3C
- Update LICENSE.md to drop the CLA part
- Update links in README.md and to mention Media WG
- Update w3c.json
- Update status, group, and links that currently target the WICG repo in the
spec itself

Also note the switch to `Level: None` in the metadata of the spec as I suspect
that the group will publish the API without any level, at least initially.

This update also drops the hyphen in `web-codecs` as discussed in:
https://github.com/WICG/web-codecs/pull/149#discussion_r595407472

The default boilerplate for the Status of This Document section of the Media WG
asserts that the specs are intended to be published as W3C Recommendations. The
text needs to be adjusted for drafts that are intended to be published as W3C
Working Group Notes.

There is no specific way to flag the targeted spec status in Bikeshed. The idea
is thus to add a `Text macro: NOTE yes` macro definition in the spec metadata
and conditionals in the boilerplate files to issue the right text.

The updates to the boilerplate files are at:
https://github.com/tabatkins/bikeshed/pull/1999

Privacy and Security Consideration sections have also been added to registry
notes that targets the relevants sections in the main WebCodecs spec.