Switch to w3c/spec-prod to automate publication to /TR
The code currently relies on Travis CI to update the gh-pages branch. This
update replaces that with the w3c/spec-prod action and enables automatic
publication of the specs as working drafts to /TR whenever a change gets pushed
to the main branch.

The update also switches to the new NOTE-ED status in Bikeshed, which makes it
possible to drop the "Text macro" workaround, see discussion in:
https://github.com/tabatkins/bikeshed/pull/1999#issuecomment-819321482