Minor changes to width/height config members.
For VideoDecoderConfig, the coded size attributes are no longer
required. They only serve as a hint for the UA to create an approriate
underlying decoder (e.g. hardware accelerated for higher resolutions)
and are not strictly required. Fixes #94.

For VideoEncoderConfig, cropWidth and cropHeight are renamed to simply
be width and height. The crop prefix was added in #113 as I realized
that we were describing the "crop" size of the encoded VideoFrames. But
I now think this risks giving the impression that the encoder is going
to crop frames. Being intuitive is more important than being pedantic.