Replace decodeMetadata() w/ tracks.ready promise.
Queue task to establish tracks upon construction. As this is no longer
user driven, we replace the method with an attribute to let users know
when the track list is "ready".

Establishing tracks was previously user driven by a call to
decodeMetadat(). After further consideration, this seems needlessly
complex. Decoding track metadata is not resource intensive and we expect
that most usage of ImageDecoder is such that they wish to decode actual
frames right away. Users who wish to defer decoding track ImageData may
still do so by deferring construction of the ImageDecoder.

This commit also includes other small fixes (formatting, output
timestamp and duration, and closing ImageDecoder if we fail to establish
tracks once data is "complete").