[Auto publication] Force job serialization
One job runs for each of the 3 documents in the repository. Each of them
updates the "gh-pages" branch. Problem is that these jobs are typically run in
parallel, leading to "not up to date" issues when "git push" commands are run.

For an example of such a problem, see:
https://github.com/w3c/webcodecs/runs/2462176606

Also see discussion in:
https://github.com/w3c/spec-prod/issues/58

This update tells GitHub to serialize the jobs. Since the jobs are relatively
independent otherwise, it also tells GitHub to run them to completion even
when one of them fails.