Drop Previous Version link from ED
Spec-prod attempts to override the "Previous version" link in the documents
before generation with Bikeshed, but this is not a real override, leading to
specs with two Previous Version links, see:
https://github.com/tabatkins/bikeshed/issues/2041

In turn, this makes publication fail when spec gets published multiple times
on the same day (because "from biblio" then reports the same URL as the
"This Version" link).

Dropping the "Previous version" link from the ED fixes the issue. The link is
not useful in an ED in any case.