Add mp4 decode/demux demo
This demo is updated and migrated from https://tguilbert-google.github.io/webcodecs/mp4/index.html.

It uses mp4box.js to demux an mp4 file and generate the decoderConfig, and VideoDecoder to decode the frames.

The VideoFrames are displayed as fast as possible on screen, without consideration for the playback rate.