Improve AudioData ctor, copyTo and close
This commit resets AudioData's internal slots when it is closed. This allows us to check if AudioData was closed by looking if audioData.numberOfFrames == 0.

It also checks for the the validity of AudioDataInit in the AudioData constructor.

It also enforces that AudioDataCopyToOptions.planeIndex is valid based off of the sample format, when calling copyTo()