Add best practices section for using WebCodecs.
Per the resolution of the chairs, WebCodecs will be exposed in
windows and worker contexts. As part of that resolution the
chairs asked that we include a best practices section.

This section recommends authors move their media pipelines to
worker contexts where possible and highlights areas of concern
for those who still choose to operate in a worker.