Publish workflow: deploy images first
Images under the "images" folder were not published to GitHub Pages, resulting
in broken links in the spec.

In addition, images that are referenced by the WebCodecs spec need to be
published to Github Pages before publication to /TR occurs, since publication
to /TR makes a copy of the image. This update makes the job that takes care
of publishing specs to /TR depend on the job that publishes the images.

The job that publishes the images is similar to the one used to deploy samples.