Make VideoColorSpace constructor argument optional
The constructor takes a `VideoColorSpaceInit` dictionary as input that does not
have any required parameter. Quoting from the Web IDL spec, the input "must be
specified as optional and have a default value provided. This is to encourage
API designs that do not require authors to pass an empty dictionary value when
they wish only to use the dictionary’s default values."
https://heycam.github.io/webidl/#idl-operations