Make VideoFrame.timestamp non nullable, less fussy
It was always intended that timestamp be required when constructing a
VideoFrame. Previously, the spec steps threw a TypeError for frames
constructed with a CanvasImageSource that didn't otherwise have a
timestamp. The only way to actually construct a frame w/ a null
timestamp was to decode an image (ImageDecoder) that was not animated.
But animated images always start at time `0` and increment by the frame
duration. Hence, `0` seems seems like a reasonable default for non-animated.

Also, for VideoFrame constructed images, rather than strictly
requiring folks to supply a timestamp when constructing with things like
Canvas, we can be less fussy and just let it default to `0`, similar to
image decoded frames. Full disclosure: This is behavior is (by mistake)
already implemented in Chrome.