Compare Document.hidden with a boolean
The spec confused the `hidden` attribute with the `visibilityState` attribute.
The latter is a string that can take the value `"hidden"`. The former is a
boolean that is true when `visibilityState` equals `"hidden"`.

This update fixes that and drops the now useless references to the
Page Visibility spec (now integrated in HTML).