Make VideoFrame.timestamp non-nullable.
Current spec always defines a timestamp for user constructed
VideoFrames. Setting a real timestamp is especially important for frames
that sent to VideoEncoder, as the timestamp is used for rate control.

With no way for users to generate a null timestamp, it makes sense to
make the type non-nullable. This also aligns with AudioData.timestamp.

This patch also addresses the one lingering opporutnity for null
timestamps: ImageDecoder used null as a default for output VideoFrames.
This patch updates that default to be zero. While it could be argued
that null is technically more correct (we don't often present images on
timeline), this is outweighed by the practical impact of using zero
which allows us to define a clear contract: VideoFrame always has a
timestamp.