Drop local custom definitions where possible
Bikeshed now contains cross-reference anchors from Webref, which means it now
knows about a number of "new" specifications, including WebCodecs, Media Capture
specs, the Infra standard, etc. See announcement at:
https://lists.w3.org/Archives/Public/spec-prod/2023JanMar/0004.html

In turn, this means that WebCodecs specs no longer need to list and maintain a
local database of cross-references, except in rare cases (such as when a term
is not exported by the referenced spec).

This update drops local definitions where ever possible, adjusting references
in the prose as needed. A few definitions remain here and there to list external
terms that are not exported, along with a couple of linking defaults in the
WebCodecs spec.

One temporary hiccup: the markup definitions of `CanvasImageSource` and
`drawImage()` are slightly incorrect. Hence the need to re-define them locally.
To be removed once the HTML spec is fixed.

In codec registration specs, note that the form
`{{VideoDecoderConfig/description|VideoDecoderConfig.description}}` is used to
output `VideoDecoderConfig.description` (with the right link to the WebCodecs
spec) when merely having `description` would be ambiguous.