Add H.265/VP8/VP9/AV1 samples to video decoding page.
This should be helpful for developer to test the browser implement
or compare the performance of different codecs on different platforms.

Note that the previous version of mp4box is not able to parse AV1
video codec string correctly, so to support AV1, we also need to
upgrade mp4box to version 0.5.2.

Safari doesn't support `gl.createShader` so change the default
renderer to `2d`.

All sample videos are encoded from `bbb-1920x1080-cfg06.mp4` using
ffmpeg and shaka packager.